requirements.txt
comfyui-frontend-package==1.21.6
comfyui-workflow-templates==0.1.25
comfyui-embedded-docs==0.2.0
torch
torchsde
torchvision
torchaudio
numpy>=1.25.0
einops
transformers>=4.28.1
tokenizers>=0.13.3
sentencepiece
safetensors>=0.4.2
aiohttp>=3.11.8
yarl>=1.18.0
pyyaml
Pillow
scipy
tqdm
psutil

websocket-client

# layerdiffuse
diffusers>=0.29.0
opencv-python

# rembg
rembg[gpu]

#non essential dependencies:
kornia>=0.7.1
spandrel
soundfile
av>=14.2.0
pydantic~=2.0

# fix for pydantic issues in cog
# https://github.com/replicate/cog/issues/1623
albumentations==1.4.3